Move Utils methods into specific helper classes
